Rename project from workload-operator to compute
This repo owns the entire compute.datumapis.com API group, but its name,
import paths, and image names still said "workload-operator" from when it
only managed the Workload resource. Renaming to "compute" lines the
project up with what it actually owns.

What this changes:
- Go import path: go.datum.net/workload-operator -> go.datum.net/compute
- Container image: ghcr.io/datum-cloud/workload-operator -> ghcr.io/datum-cloud/compute
- Install bundle: ghcr.io/datum-cloud/workload-operator-kustomize ->
  ghcr.io/datum-cloud/compute-kustomize

What does NOT change:
- The compute.datumapis.com API group (Workload, WorkloadDeployment,
  Instance) - existing resources keep working with no migration.
- Leader election lease key, so there are no orphaned leases on cutover.
