+++ internal/controller/instancecontrol/stateful/stateful_control_util_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"go.datum.net/workload-operator/api/v1alpha"
-	"go.datum.net/workload-operator/internal/controller/instancecontrol"
+	"go.datum.net/compute/api/v1alpha"
+	"go.datum.net/compute/internal/controller/instancecontrol"
 )
 
 func TestGetInstanceOrdinal(t *testing.T) {
