+++ internal/controller/workloaddeployment_controller.go
@@ -26,11 +26,11 @@ import (
 	mcmanager "sigs.k8s.io/multicluster-runtime/pkg/manager"
 	mcreconcile "sigs.k8s.io/multicluster-runtime/pkg/reconcile"
 
+	computev1alpha "go.datum.net/compute/api/v1alpha"
 	networkingv1alpha "go.datum.net/network-services-operator/api/v1alpha"
-	computev1alpha "go.datum.net/workload-operator/api/v1alpha"
 
-	"go.datum.net/workload-operator/internal/controller/instancecontrol"
-	instancecontrolstateful "go.datum.net/workload-operator/internal/controller/instancecontrol/stateful"
+	"go.datum.net/compute/internal/controller/instancecontrol"
+	instancecontrolstateful "go.datum.net/compute/internal/controller/instancecontrol/stateful"
 )
 
 // WorkloadDeploymentReconciler reconciles a WorkloadDeployment object
